Create README
Hole deviation of a drilling well is the unintentional departure of the drill bit from a planned wellbore trajectory. Whether it involves drilling a straight straight section, or a curved hole section, the deviation of drill bit from the desired path can lead to drilling problems such as higher non productive time, more drilling costs and also the legal issues if deviation is extreme and enters in non leased block, so while drilling frequent directional surveys are conducted to make sure that deviation is minim. This algorithm might be helpful to identify the deviation as drilling progresses.